pop up

+++ src/app/login/login.component.ts
@@ -1,3 +1,4 @@
+
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
@@ -5,6 +6,7 @@ import { RegisterComponent } from '../register/register.component';
 import { LoginDataModel } from '../shared/models/loginData.model';
 import { HttpClient } from '@angular/common/http';
 import { APP_CONFIG } from '../shared/constants/constants';
+import Swal from 'sweetalert2'; 
 
 @Component({
   selector: 'app-login',
@@ -17,12 +19,10 @@ export class LoginComponent implements OnInit {
   public loginModel!: LoginDataModel;
 
   ngOnInit(): void {
-    this.loginForm = new FormGroup(
-      {
-        email: new FormControl('', [Validators.required, Validators.email]),
-        password: new FormControl('', [Validators.required, Validators.minLength(6)])
-      }
-    );
+    this.loginForm = new FormGroup({
+      email: new FormControl('', [Validators.required, Validators.email]),
+      password: new FormControl('', [Validators.required, Validators.minLength(6)])
+    });
   }
 
   constructor(private _dilog: MatDialog, private http: HttpClient) { }
@@ -30,43 +30,62 @@ export class LoginComponent implements OnInit {
   openRegisterEmForm() {
     const dialogRef = this._dilog.open(RegisterComponent, {
       width: '600px',
-    })
-
+    });
   }
 
   hide = true;
-  // get emailInput() { return this.loginForm.get('email'); }
-  // get passwordInput() { return this.loginForm.get('password'); }  
 
   onLogin() {
     const obj = {
       email_user: this.loginForm.value.email,
       password_user: this.loginForm.value.password,
-    }
+    };
 
-    console.log(obj);
+    this.http.post(this.urlbackend + '/api/v1/login', obj).subscribe({
+      next: (res: any) => {
+        if (res.id_user == null) {
+          
+          Swal.fire({
+            title: 'ไม่พบบัญชีผู้ใช้!',
+            text: 'ไม่พบอีเมลนี้ในระบบ กรุณาตรวจสอบอีกครั้ง',
+            icon: 'warning',
+            confirmButtonText: 'ตกลง'
+          });
+          sessionStorage.clear();
 
-    this.http.post(this.urlbackend+'/api/v1/login', obj).subscribe((res: any) => {
-      console.log(res);
+        } else if (res.status_login) {
+          
+          Swal.fire({
+            title: 'เข้าสู่ระบบสำเร็จ!',
+            text: 'ยินดีต้อนรับคุณ ' + (res.name_user ? res.name_user : 'ผู้ใช้งาน'),
+            icon: 'success',
+            timer: 1500,
+            showConfirmButton: false
+          }).then(() => {
+            sessionStorage.setItem('user', JSON.stringify(res));
+            window.location.reload();
+          });
 
-
-      if (res.id_user == null) {
-        console.log("ไม่มีอีเมล");
-        sessionStorage.clear()
-
-      } else if (res.status_login) {
-        console.log("login สำเร็จ");
-        sessionStorage.setItem('user', JSON.stringify(res))
-        window.location.reload()
-      } else {
-        console.log("รหัสผ่านไม่่ถูกต้อง");
-        sessionStorage.clear()
+        } else {
+      
+          Swal.fire({
+            title: 'รหัสผ่านผิด!',
+            text: 'กรุณาตรวจสอบรหัสผ่านของคุณใหม่อีกครั้ง',
+            icon: 'error',
+            confirmButtonText: 'ลองใหม่'
+          });
+          sessionStorage.clear();
+        }
+      },
+      error: (err) => {
+    
+        Swal.fire({
+          title: 'เกิดข้อผิดพลาด!',
+          text: 'ไม่สามารถติดต่อเซิร์ฟเวอร์ได้ในขณะนี้',
+          icon: 'error',
+          confirmButtonText: 'ตกลง'
+        });
       }
-
-    })
-
-    console.log(this.loginModel);
+    });
   }
 }
-  
-
